refactor: transform LanguageTOggle into a hook
